feat: update sig prover to support withdraw tx
