+++ crates/signature-prover/server/src/keystore_types/contract.rs
@@ -1,0 +1,9 @@
+use alloy_sol_types::sol;
+
+sol!(
+    #[derive(Debug)]
+    #[sol(rpc)]
+    #[allow(clippy::too_many_arguments)]
+    AxiomKeystoreRollup,
+    "src/keystore_types/abi/AxiomKeystoreRollup.json"
+);
